+++ src-docs/src/views/accordion/accordion_form.js
@@ -51,7 +51,7 @@ const buttonContent = (
       </EuiFlexItem>
 
       <EuiFlexItem>
-        <EuiTitle size="s" className="euiAccordionForm__title">
+        <EuiTitle size="xs">
           <h3>Webhook</h3>
         </EuiTitle>
       </EuiFlexItem>
@@ -78,12 +78,6 @@ const extraAction = (
 
 export default () => (
   <div>
-    <EuiTitle size="s">
-      <h3>I am a complicated, highly styled, repeatable form!</h3>
-    </EuiTitle>
-
-    <EuiSpacer size="l" />
-
     <EuiAccordion
       id="accordionForm1"
       className="euiAccordionForm"