+++ src/components/panel/split_panel/split_panel.tsx
@@ -19,20 +19,12 @@
 
 import React, { FunctionComponent, ReactNode, HTMLAttributes } from 'react';
 import classNames from 'classnames';
-import { EuiPanel, _EuiPanelProps, EuiPanelProps } from '../panel';
+import { EuiPanel, _EuiPanelProps } from '../panel';
 import { EuiBreakpointSize } from '../../../services/breakpoint';
 import { useIsWithinBreakpoints } from '../../../services/hooks';
 
 export type _EuiSplitPanelInnerProps = HTMLAttributes<HTMLDivElement> &
-  Omit<
-    _EuiPanelProps,
-    | 'hasShadow'
-    | 'hasBorder'
-    | 'borderRadius'
-    | 'betaBadgeLabel'
-    | 'betaBadgeTooltipContent'
-    | 'betaBadgeTitle'
-  >;
+  Omit<_EuiPanelProps, 'hasShadow' | 'hasBorder' | 'borderRadius'>;
 
 /**
  * Consumed via `EuiSplitPanel.Inner`.
@@ -53,7 +45,11 @@ export const _EuiSplitPanelInner: FunctionComponent<_EuiSplitPanelInnerProps> = 
   };
 
   return (
-    <EuiPanel className={classes} {...panelProps} {...(rest as EuiPanelProps)}>
+    <EuiPanel
+      element="div"
+      className={classes}
+      {...panelProps}
+      {...(rest as _EuiPanelProps)}>
       {children}
     </EuiPanel>
   );
@@ -73,13 +69,7 @@ export type _EuiSplitPanelOuterProps = HTMLAttributes<HTMLDivElement> & {
    * Remove completely with `false` or provide your own list of breakpoint sizes to stack on.
    */
   responsive?: false | EuiBreakpointSize[];
-} & Omit<
-    _EuiPanelProps,
-    | 'paddingSize'
-    | 'betaBadgeLabel'
-    | 'betaBadgeTooltipContent'
-    | 'betaBadgeTitle'
-  >;
+} & Omit<_EuiPanelProps, 'paddingSize'>;
 
 /**
  * Consumed via `EuiSplitPanel.Outer`.
@@ -111,7 +101,7 @@ export const _EuiSplitPanelOuter: FunctionComponent<_EuiSplitPanelOuterProps> = 
       paddingSize="none"
       grow={false}
       className={classes}
-      {...(rest as EuiPanelProps)}>
+      {...(rest as _EuiPanelProps)}>
       {children}
     </EuiPanel>
   );